+++ packages/release-tools/createRelease.ts
@@ -1,51 +1,70 @@
-import { join } from "path";
+#! /usr/bin/env -S node --experimental-strip-types
+import { join } from "node:path";
 import {
     addRemoteWithAuthentication,
     execShellCommand,
     getPackageInfo,
     gh,
-    PackageInfo,
+    type PackageInfo,
     ChangelogFileWrapper
-} from "./utils";
+} from "./utils/index.ts";
+import { ArgumentError, handleError } from "./utils/errors.ts";
+import { format } from "node:util";
+
+const packages = new Map([
+    [
+        "pluggable-widgets-tools",
+        {
+            name: "pluggable-widgets-tools",
+            fullName: "Pluggable Widgets Tools"
+        }
+    ],
+    [
+        "generator-widget",
+        {
+            name: "generator-widget",
+            fullName: "Pluggable Widgets Generator"
+        }
+    ]
+]);
 
 main().catch(e => {
-    console.error(e);
-    process.exit(-1);
+    handleError(
+        e,
+        format(
+            "\nUsage: %s <package-name> <branch-name>\nRequired environment variables: GH_NAME, GH_EMAIL, GH_USERNAME, GH_PAT",
+            process.argv[1]
+        )
+    );
+    process.exit(1);
 });
 
 async function main(): Promise<void> {
-    const packages = new Map([
-        [
-            "pwt",
-            {
-                name: "pluggable-widgets-tools",
-                fullName: "Pluggable Widgets Tools"
-            }
-        ],
-        [
-            "gw",
-            {
-                name: "generator-widget",
-                fullName: "Pluggable Widgets Generator"
-            }
-        ]
-    ]);
     const arg = process.argv[2];
     const mendixPackage = packages.get(arg);
-    if (!mendixPackage) throw new Error(`Argument "${arg}" is not a valid package name`);
+    if (!mendixPackage)
+        throw new ArgumentError(format("package name of (%s)", packages.keys().toArray().join(", ")), arg);
+    const branch = process.argv[3];
+    if (!branch || !/^(master|version\/.*)$/.test(branch))
+        throw new ArgumentError("branch name of (master, version/*)", branch);
 
-    const pwtPath = join(__dirname, "../", mendixPackage.name);
+    const dirname = new URL("./", import.meta.url).pathname;
+    const pwtPath = join(dirname, "../", mendixPackage.name);
 
     // 1. Get release info
-    console.log(`Getting the release information for ${mendixPackage.name}...`);
-    console.log(`directory:`, pwtPath);
+    console.log("Gathering release information");
+    console.log("  Package:   %s", mendixPackage.name);
+    console.log("  Directory: %s", pwtPath);
+    console.log("  Branch:    %s", branch);
 
     const packageInfo = await getPackageInfo(pwtPath);
     packageInfo.packageName = mendixPackage.name;
     packageInfo.packageFullName = mendixPackage.fullName;
     const releaseTag = `${packageInfo.packageName}-v${packageInfo.version.format()}`;
-    const changelog = ChangelogFileWrapper.fromFile(`${pwtPath}/CHANGELOG.md`);
 
+    console.log("  Version:   %s", packageInfo.version.format());
+
+    const changelog = ChangelogFileWrapper.fromFile(`${pwtPath}/CHANGELOG.md`);
     // 2. Check prerequisites
     // 2.1. Check if current version is already in CHANGELOG
     if (changelog.hasVersion(packageInfo.version)) {
@@ -77,7 +96,7 @@ async function main(): Promise<void> {
 
     // 4.2 Update CHANGELOG.md and create PR
     console.log("Creating PR with updated CHANGELOG.md file...");
-    await updateChangelogsAndCreatePR(packageInfo, changelog, releaseTag, remoteName);
+    await updateChangelogsAndCreatePR(packageInfo, changelog, releaseTag, remoteName, branch);
 
     // 4.3 Create release
     console.log("Creating Github release...");
@@ -99,7 +118,8 @@ async function updateChangelogsAndCreatePR(
     packageInfo: PackageInfo,
     changelog: ChangelogFileWrapper,
     releaseTag: string,
-    remoteName: string
+    remoteName: string,
+    branch: string
 ): Promise<void> {
     const releaseBranchName = `${releaseTag}-update-changelog`;
 
@@ -121,7 +141,7 @@ async function updateChangelogsAndCreatePR(
     await gh.createGithubPRFrom({
         title: `${packageInfo.packageFullName} v${packageInfo.version.format()}: Update changelog`,
         body: "This is an automated PR that merges changelog update to master.",
-        base: "master",
+        base: branch,
         head: releaseBranchName,
         repo: packageInfo.repositoryUrl
     });