+++ packages/release-tools/utils/changelog-parser/cli.ts
@@ -1,22 +1,46 @@
-import { ChangelogFileWrapper } from ".";
-import { resolve } from "path";
-import * as process from "process";
+#!/usr/bin/env -S node --experimental-strip-types
+import { ChangelogFileWrapper } from "./index.ts";
+import { resolve } from "node:path";
+import * as process from "node:process";
+import { ArgumentError, handleError } from "../errors.ts";
+import { format } from "node:util";
 
-function reformat(filePath: string): void {
-    try {
-        const changelog = ChangelogFileWrapper.fromFile(filePath);
-
-        changelog.save();
-    } catch (e) {
-        console.log(e);
-        process.exit();
+function ensureFilePath(filePath: string | undefined) {
+    if (!filePath) {
+        throw new ArgumentError("filepath", filePath);
     }
+    return resolve(filePath);
+}
+
+function reformat(filePath: string | undefined): void {
+    filePath = ensureFilePath(filePath);
+    const changelog = ChangelogFileWrapper.fromFile(filePath);
+    changelog.save();
+    console.log("Finished formatting %s", filePath);
+}
+
+function parse(filePath: string | undefined): void {
+    filePath = ensureFilePath(filePath);
+    ChangelogFileWrapper.fromFile(filePath);
+    console.log("Successfully parsed %s", filePath);
 }
 
 function main(): void {
-    switch (process.argv[2]) {
-        case "reformat":
-            return reformat(resolve(process.argv[3]));
+    try {
+        switch (process.argv[2]) {
+            case "reformat":
+                return reformat(process.argv[3]);
+            case "parse":
+                return parse(process.argv[3]);
+            default:
+                throw new ArgumentError("command of reformat", process.argv[2]);
+        }
+    } catch (e) {
+        handleError(
+            e,
+            format("Usage: %s <command>\n\nCommands:\n\treformat <filePath>\n\tparse <filePath>", process.argv[1])
+        );
+        process.exit(1);
     }
 }